langgraph/langserve/pyproject.toml
[project]
name = "langserve-langgraph-demo"
version = "0.1.0"
description = "4 LangGraph agent examples served via LangServe"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0,<0.4.0",
    "langchain-openai>=0.3.0,<0.4.0",
    "langchain-core>=0.3.0,<0.4.0",
    "langserve[all]>=0.3.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
